add Telegram notifications and make structure for functions and notifications
